Add more typings
