+++ game-frontend/src/components/canvas-objects/Food.tsx
@@ -1,7 +1,12 @@
 import React from 'react';
-import Circle from '../pixi/circle';
+import Circle from '../pixi/Circle';
+import Position from '../../../../darwin-types/Position';
 
-function Unit({ position }) {
+type Props = {
+  position: Position;
+};
+
+function Unit({ position }: Props): JSX.Element {
   return <Circle position={position} radius={2} color={0xff700b} />;
 }
