+++ game-frontend/src/components/canvas-objects/Unit.tsx
@@ -1,7 +1,12 @@
 import React from 'react';
-import Square from '../pixi/square';
+import Square from '../pixi/Square';
+import Position from '../../../../darwin-types/Position';
 
-function Unit({ position }) {
+type Props = {
+  position: Position;
+};
+
+function Unit({ position }: Props): JSX.Element {
   return <Square position={position} size={5} color={0xffd900} />;
 }
