+++ game-frontend/src/components/pixi/Circle.tsx
@@ -1,26 +1,30 @@
 import React from 'react';
 import { Graphics } from '@inlet/react-pixi';
+import { PixiGeometricFormProps } from './PixiGeometricFormProps';
 
-function Square({
-  size,
+interface Props extends PixiGeometricFormProps {
+  radius: number;
+}
+
+function Circle({
+  radius,
   position,
   color = 0xffffff,
   alpha = 1,
   lineWidth = 1,
-}) {
+}: Props): JSX.Element {
   const { x, y } = position;
   return (
     <Graphics
       draw={g => {
         // clear the graphics
         g.clear();
-
+        // start drawing
         g.lineStyle(lineWidth, color, alpha);
-
-        g.drawRect(x, y, size, size);
+        g.drawCircle(x, y, radius);
       }}
     />
   );
 }
 
-export default Square;
+export default Circle;
