+++ game-frontend/src/components/pixi/PixiGeometricFormProps.ts
@@ -1,0 +1,8 @@
+import Position from '../../../../darwin-types/Position';
+
+export type PixiGeometricFormProps = {
+  position: Position;
+  alpha?: number;
+  color?: number;
+  lineWidth?: number;
+};
