+++ game-frontend/src/components/pixi/Square.tsx
@@ -1,25 +1,31 @@
 import React from 'react';
 import { Graphics } from '@inlet/react-pixi';
+import { PixiGeometricFormProps } from './PixiGeometricFormProps';
 
-function Circle({
-  radius,
+interface Props extends PixiGeometricFormProps {
+  size: number;
+}
+
+function Square({
+  size,
   position,
   color = 0xffffff,
   alpha = 1,
   lineWidth = 1,
-}) {
+}: Props): JSX.Element {
   const { x, y } = position;
   return (
     <Graphics
       draw={g => {
         // clear the graphics
         g.clear();
-        // start drawing
+
         g.lineStyle(lineWidth, color, alpha);
-        g.drawCircle(x, y, radius);
+
+        g.drawRect(x, y, size, size);
       }}
     />
   );
 }
 
-export default Circle;
+export default Square;
